requirements.txt
nanodlna
gTTS
Pillow